fazendo alguns exercicios em python

+++ mundo01-python/Tratando dados e Fazendo Contas/Operadores Aritmeticos/desafio08.py
@@ -1,4 +1,4 @@
-n1 = int(input('Primeira Nota: '))
-n2 = int(input('Segunda Nota: '))
+n1 = float(input('Primeira Nota: '))
+n2 = float(input('Segunda Nota: '))
 media = (n1 + n2) / 2
 print('A media do aluno e de {}'.format(media))
